+++ core/src/main/java/discord4j/core/object/presence/Activity.java
@@ -24,7 +24,7 @@ import reactor.util.annotation.Nullable;
 
 import java.time.Instant;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 public class Activity {
 
@@ -149,13 +149,13 @@ public class Activity {
      *
      * @return The party's current size, if present.
      */
-    public OptionalInt getCurrentPartySize() {
+    public OptionalLong getCurrentPartySize() {
         if (richData == null) {
-            return OptionalInt.empty();
+            return OptionalLong.empty();
         }
 
-        final Integer currentPartySize = richData.getCurrentPartySize();
-        return (currentPartySize == null) ? OptionalInt.empty() : OptionalInt.of(currentPartySize);
+        final Long currentPartySize = richData.getCurrentPartySize();
+        return (currentPartySize == null) ? OptionalLong.empty() : OptionalLong.of(currentPartySize);
     }
 
     /**
@@ -163,13 +163,13 @@ public class Activity {
      *
      * @return The party's max size, if present.
      */
-    public OptionalInt getMaxPartySize() {
+    public OptionalLong getMaxPartySize() {
         if (richData == null) {
-            return OptionalInt.empty();
+            return OptionalLong.empty();
         }
 
-        final Integer maxPartySize = richData.getMaxPartySize();
-        return (maxPartySize == null) ? OptionalInt.empty() : OptionalInt.of(maxPartySize);
+        final Long maxPartySize = richData.getMaxPartySize();
+        return (maxPartySize == null) ? OptionalLong.empty() : OptionalLong.of(maxPartySize);
     }
 
     /**
